+++ R/calculateEvents.R
@@ -5,6 +5,7 @@
 #' genotypes of the trio as input and includes a final step to simplify the
 #' results into blocks.
 #'
+<<<<<<< HEAD
 #' @param largeCollapsedVcf The VCF file in the general format
 #' @param hmm Hidden Markov Model used to infer the events. The format should 
 #' adhere to the general HMM format with a series of elements:
@@ -16,11 +17,18 @@
 #' observations in the "emissionProbs" matrix.
 #' @param genotypes Possible GT formats and its correspondence with the hmm
 #' (largeCollapsedVcf) with VariantAnnotation package.
+=======
+#' @param largecollapsedVcf The VCF file in the general format
+#' @param hmm hmm constructed for identifying hidden states
+#' @param genotypes Possible GT formats and its correspondency with the hmm
+#' (largecollapsedVcf) with VariantAnnotation package.
+>>>>>>> upstream/devel
 #' @return Dataframe object containing blocks of predicted events.
 #' @export
 #' @examples
 #' file <- system.file(package = "UPDhmm", "extdata", "test_het_mat.vcf.gz")
 #' vcf <- VariantAnnotation::readVcf(file)
+<<<<<<< HEAD
 #' processedVcf <- vcfCheck(vcf,
 #'     proband = "NA19675", mother = "NA19678",
 #'     father = "NA19679"
@@ -246,3 +254,57 @@ calculateEvents <-
 
         return(block_def)
     }
+=======
+#' processedVcf <- vcfCheck(vcf,proband = "NA19675",mother = "NA19678",
+#' father = "NA19679")
+#'
+#' calculateEvents(processedVcf)
+calculateEvents <-
+function(largecollapsedVcf = NULL,hmm = "hmm", genotypes = NULL) {
+    if (!methods::is(hmm, "list"))
+{utils::data("hmm", package = "UPDhmm", envir = environment())}
+if (base::is.null(genotypes))
+{genotypes <- c("0/0" = "1", "0/1" = "2", "1/0" = "2", "1/1" = "3",
+    "0|0" = "1", "0|1" = "2", "1|0" = "2", "1|1" = "3")}
+else {genotypes}
+# 1 split the vcf into chromosomes
+split_vcf_raw <-
+    split(largecollapsedVcf,f = GenomicRanges::seqnames(largecollapsedVcf))
+#2 apply viterbi (# Return NULL if an error occurs)
+split_vcf <- lapply(split_vcf_raw, function(x)
+{tryCatch( applyViterbi(
+    largecollapsedVcf = x, hmm = hmm, genotypes = genotypes),
+    error = function(e) NULL)})
+#3 as_df
+split_vcf_df <- lapply(split_vcf, function(x)
+{tryCatch(asDfVcf(
+    largecollapsedVcf=x,genotypes = genotypes),
+    error = function(e) NULL) })
+#4 Create blocks of contiguous positions with same state
+blocks_state <- lapply(split_vcf_df, function(df) {
+tryCatch(blocksVcf(df),error = function(e) NULL)})
+
+#5 simplify all chr objects into one data.frame
+def_blocks_states <- base::Reduce(rbind,blocks_state)
+
+#6 Filter normal state blocks , sexual chromosomes and isolated variants
+filtered_def_blocks_states <-
+def_blocks_states[def_blocks_states$n_snps > 1
+    & def_blocks_states$group != "normal" &
+    !(def_blocks_states$seqnames %in% c("chrX", "X")), ]
+
+# 7 Calculate statistics parameters
+if (nrow(filtered_def_blocks_states) > 0) {
+blocks_list <- lapply(seq_len(nrow(filtered_def_blocks_states)),
+    function(i){addOr(
+    filtered_def_blocks_states = filtered_def_blocks_states[i, , drop = FALSE],
+    largecollapsedVcf = largecollapsedVcf,
+    hmm = hmm,
+    genotypes = genotypes)})}
+else{block_def<-data.frame()}
+
+# 8 Transform final output to data.frame
+block_def <- base::Reduce(rbind,blocks_list)
+return(block_def)
+}
+>>>>>>> upstream/devel