+++ NEWS.md
@@ -1,5 +1,6 @@
 # UPDhmm 0.99.0
 
+<<<<<<< HEAD
 First release of UPDhmm package
 
 # UPDhmm 0.99.1
@@ -9,3 +10,6 @@ First pre-review corrections
 # UPDhmm 0.99.2
 
 Correction of reviewer's suggestions
+=======
+* Added a `NEWS.md` file to track changes to the package.
+>>>>>>> upstream/devel